Create autoCopyCatTrainingFramework_public_v001.py
If anyone is interested:

I have written a tool that allows for semi-procedural Copy Cat training in Nuke by pulling an arbitrary number of random images from plates in a job directory at a facility.

Having so many sub-functions may have been a clunky approach but the goal was to begin exploring a framework for procedural training and I believe it makes the tool more modular and adaptable to different training purposes and pipelines.  

With this tool you can create 'num_iterations' of read nodes to quickly and procedurally generate a random data set to then process and train on with Copy Cat.

The main 'get_random_image_path' function takes a 'jobs' path and returns a path to a single random image from an image sequence. Currently it is tailored to return an image from a plate in a random job/sequence/shot but could be adjusted to suit other needs. This function is going to be the one that requires the most leg work to point somewhere else, but I believe it is well suited for the job as it is semi-path-agnostic and the pathing it is returning at a given point is well documented (also if you give GPT this function and a sample file path from what you are trying to access it is very likely to succeed at modifying it for you).

From there 'create_read_node_with_image' defines the read node and knob values that need created. There is a check to flip certain job directories' exrs to use a rec709 color transform instead of the default linear, this is because some facilities will write graded plates out with that transform baked in. 

Next, 'create_copy_node' defines a 'copy' node. This tool is written assuming you want to train against processing happening inside of Nuke. It is copying a processed set of channels into the main read node rgb stream in a processed.rgb layer set. (This is because the CopyCat node works on image sequences and these read nodes will all later be appended to an image sequence. This was the easiest way to make sure the destination and source images would be aligned in the sequence as you can now just shuffle out the processed chain after the appendClip).

Next, 'create_and_connect_nodes' does two things. First it is where you would place the processing you want to train against. In this case I am downsizing the image and clamping the brightest parts of the image at .7 (I am working on a tool that would allow you to select nodes in a script and automatically copy them to the formatting you would need to paste directly into this section). Second it connects these nodes in sequence and creates the copy node from the above function.

Next, 'create_append_clip_node' iterates for the given number of read nodes and plugs them all into an appendClip node to create an image sequence. It also shuffles the processed channels back to rgba, removes all extra channels, and connects the two streams to a copyCat node. 

At the end, create_append_clip_node is called with all the necessary arguments and variables.